requirements.txt
pinecone-client==4.1.0
openai==1.42.0
pandas==2.2.2
fastapi==0.112.2
uvicorn==0.30.6
redis==5.0.8
pydantic==2.8.2
